Moved snippet files
