+++ ftp/src/test/java/docs/javadsl/FtpSettingsExample.java
@@ -2,7 +2,7 @@
  * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package akka.stream.alpakka.ftp.examples;
+package docs.javadsl;
 
 // #create-settings
 import akka.stream.alpakka.ftp.FtpCredentials;
